# server/sched.go:  Added support for grouping GPUs ( Power-Consumption + Low-PCIe Speed Performance gains)