@igs-raiderler 2025-05-12
Interesting solution to this problem! I've been working on similar challenges recently.