@dan-and 2025-06-05
@rick-github , thanks for pointing me to the VRAM allocation effect of my patch, as this is next to the power-consumption is a huge benefit. 

There are examples of how much VRAM overhead ollama is consuming, when we are using the tactic of spreading the model over all available GPUs instead of grouping them.