@dan-and 2025-06-23
Hi @jmorganca and @mxyng 

Do you see major issues in my patch, which may prevent including it into the master branch? If so, please give me feedback so that I can update accordingly 

Kind regards