@dan-and 2025-07-09
Hi Jesse,
thanks for the feedback. 
I will rearrange my patch so that it is the default behavior. Give me 2-3 days and it will be updated.