@dan-and 2025-07-11
@jessegross  
Updated to make it the default behavior, and OLLAMA_SCHED_SPREAD=1 will spread it evenly.

Because of my patch, I have introduced two slog.Info and one slog.Debug into the code. It helped me quite well to figure out if ollama is doing the correct allocation with various system setups and configs. Should I keep them in the code, or would you suggest  removing them or moving the slog.Info to slog.Debug to keep the amount of logs low? 

Regards,
Daniel 

PS: It works great on my 7 GPU system:  

OLLAMA_SCHED_SPREAD=1  

> daniel@gpu:~/source/ollama_dan$ ollama run qwen3:4b --verbose "say hi!" 
> Thinking...
> Okay, the user said "say hi!" so I need to respond with a friendly greeting. Let me make sure to keep it warm and welcoming. Maybe add an emoji to make it more lively. I should also ask how I can assist them today. Let me 
> check the tone—should be positive and eager to help. Alright, that should do it.
> ...done thinking.
> 
> Hello! 😊 How can I assist you today?
> 
> total duration:       11.976054545s
> load duration:        **10.255610156s**
> prompt eval count:    13 token(s)
> prompt eval duration: 358.652306ms
> prompt eval rate:     36.25 tokens/s
> eval count:           88 token(s)
> eval duration:        1.36080978s
> eval rate:            **64.67 tokens/s**
> daniel@gpu:~/source/ollama_dan$ ollama ps
> NAME        ID              SIZE     PROCESSOR    CONTEXT    UNTIL               
> qwen3:4b    2bfd38a7daaf    **20 GB**    100% GPU     8192       29 minutes from now    
> 


OLLAMA_SCHED_SPREAD=0 or not set at all

> daniel@gpu:~/source/ollama_dan$ ollama run qwen3:4b --verbose "say hi!"
> Thinking...
> Okay, the user said "say hi!" so I need to respond in a friendly and welcoming manner. Let me start by acknowledging their greeting. Maybe use an emoji to keep it light and approachable.
> I should mention my name, Qwen, to establish identity. Then, offer assistance by asking how I can help them today. Keep it concise but open-ended so they feel comfortable to share their
> needs. Make sure the tone is positive and eager to help. Let me put that all together.
> ...done thinking.
> 
> Hello! I'm Qwen, your helpful assistant. How can I assist you today? 😊
> 
> total duration:       5.474187439s
> load duration:        **3.6653167s**
> prompt eval count:    13 token(s)
> prompt eval duration: 139.89265ms
> prompt eval rate:     92.93 tokens/s
> eval count:           125 token(s)
> eval duration:        1.668211081s
> eval rate:            **74.93 tokens/s**
> daniel@gpu:~/source/ollama_dan$ ollama ps
> NAME        ID              SIZE      PROCESSOR    CONTEXT    UNTIL
> qwen3:4b    2bfd38a7daaf    **7.3 GB**    100% GPU     8192       29 minutes from now