@jessegross 2025-07-12
Yes, I think you can remove the extra logging and the memory calculations that go into them. Please also retain existing comments and TODOs that aren't affected by your change.

Since you are iterating from smallest to largest number of GPUs, once you find a set that fits, you can just return that. You don't need to keep looking and find the best one - no later tries are going to use fewer GPUs.

I have some in-progress code that does something similar for a different code path, you can use it as a reference:
https://github.com/ollama/ollama/blob/f4c2e622425e4f987f664e16f60a62b0a4c4e070/llm/server.go#L777