@dan-and 2025-07-25
@jessegross  any additional comments and hints for me?